backend/requirements.txt
# Django Framework
Django==5.2.7

# Django REST Framework
djangorestframework==3.15.2

# CORS and Filtering
django-cors-headers==4.6.0
django-filter==24.3

# Environment Variables
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.10

# Production WSGI Server
gunicorn==23.0.0

# Static Files Serving
whitenoise==6.8.2

# Image Processing & QR Codes
Pillow==11.0.0
qrcode[pil]==8.0

# PDF Generation
reportlab==4.2.5

# Security & Performance (optional but recommended)
django-ratelimit==4.1.0